CHANGED: Return condition before var inits in Log.cpp + colorcode type to uint8 due to compile warnings CHANGED: Use proper format specifiers in output (%d is for signed int)

+++ src/arcemu-shared/Log.h
@@ -64,7 +64,7 @@ public:
   void SetFileLoggingLevel(int32 level);
   void SetScreenLoggingLevel(int32 level);
 
-  void outColor(uint32 colorcode, const char * str, ...);
+  void outColor(uint8 colorcode, const char * str, ...);
   
 #ifdef WIN32
   HANDLE stdout_handle, stderr_handle;
